Moved client functions tests to new file
