+++ test/30_debate/debate.js
@@ -19,165 +19,6 @@ describe('Debate test', () => {
         await debateManager.start();
     });
 
-    describe("Debate client functions", () => {
-        let client;
-        let id;
-        let debate;
-
-        before((done) => {
-            admin = io.connect(`http://localhost:${SocketConfig.SOCKET_PORT}${SocketConfig.PRIVILEGED_NAMESPACE}`, {
-                path: SocketConfig.DEFAULT_PATH,
-                forceNew: true,
-                query: {
-                    password: `${SocketConfig.ADMIN_PASSWORD}`,
-                    username: `admin`
-                }
-            });
-
-            let debateInfo = {
-                title: 'My new debate',
-                description: 'Test debate'
-            };
-            admin.emit("newDebate", debateInfo, (debateID) => {
-                id = debateID;
-                done();
-            });
-        });
-
-        beforeEach((done) => {
-            debate = debateManager.nspAdmin.getActiveDebate(id);
-
-            client = io.connect(`http://localhost:${SocketConfig.SOCKET_PORT}${SocketConfig.DEBATE_NAMESPACE_PREFIX}${id}`, {
-                path: SocketConfig.DEFAULT_PATH,
-                forceNew: true,
-                query: {
-                    uuid: '2345675432'
-                }
-            });
-
-            client.on('connect', () => {
-                done();
-            });
-        });
-
-        describe('getQuestions', () => {
-            it('no questions', (done) => {
-                client.emit('getQuestions', (questions) => {
-                    questions.length.should.equal(0);
-                    done();
-                });
-            });
-
-            it('available questions', (done) => {
-                const NB_QUESTIONS = 3;
-                for (let i = 0; i < NB_QUESTIONS; ++i)
-                    debate.sendNewQuestion(new debate.Question(`Question${i}`, ['...']));
-
-                client.emit('getQuestions', (questions) => {
-                    questions.length.should.equal(NB_QUESTIONS);
-                    for (let i = 0; i < questions.length; ++i)
-                        questions[i].title.should.equal(`Question${i}`);
-
-                    done();
-                })
-            });
-        });
-
-        describe('newQuestion', () => {
-            it('valid question', (done) => {
-                client.on('newQuestion', (questionObj) => {
-                    questionObj.title.should.equal('Does this test work ?');
-                    questionObj.answers[0].should.equal('Yes');
-                    questionObj.answers[1].should.equal('No');
-                    done();
-                });
-
-                debate.sendNewQuestion(new debate.Question('Does this test work ?', ['Yes', 'No']));
-            });
-        });
-
-        describe('answerQuestion', () => {
-            it('valid response', (done) => {
-                client.on('newQuestion', (questionObj) => {
-                    client.emit('answerQuestion', {questionId : questionObj.id, answerId : 0}, (res) => {
-                        res.should.equal(true);
-                        done();
-                    });
-                });
-
-                debate.sendNewQuestion(new debate.Question('Does this test work ?', ['Yes', 'No']));
-            });
-
-            it('invalid questionID', (done) => {
-                client.on('newQuestion', (questionObj) => {
-                    client.emit('answerQuestion', {questionId : -1, answerId : 1}, (res) => {
-                        res.should.equal(false);
-                        done();
-                    });
-                });
-
-                debate.sendNewQuestion(new debate.Question('Does this test work ?', ['Yes', 'No']));
-            });
-
-            it('invalid object', (done) => {
-                client.on('newQuestion', (questionObj) => {
-                    client.emit('answerQuestion', {myFieldIsInvalid: 12}, (res) => {
-                        res.should.equal(false);
-                        done();
-                    });
-                });
-
-                debate.sendNewQuestion(new debate.Question('Does this test work ?', ['Yes', 'No']));
-            });
-        });
-
-        describe('answerOpenQuestion', () => {
-            it('valid response', (done) => {
-                client.on('newQuestion', (questionObj) => {
-                    client.emit('answerOpenQuestion', {questionId : questionObj.id, answer : 'Hopefully, yes'}, (res) => {
-                        res.should.equal(true);
-
-                        let question = debate.questions.get(questionObj.id);
-                        question.answers[0].answer.should.equal('Hopefully, yes');
-                        done();
-                    });
-                });
-
-                debate.sendNewQuestion(new debate.Question('Does this test work ?', null, true));
-            });
-
-            it('invalid questionID', (done) => {
-                client.on('newQuestion', (questionObj) => {
-                    client.emit('answerOpenQuestion', {questionId : -1, answer : 'Hey'}, (res) => {
-                        res.should.equal(false);
-                        done();
-                    });
-                });
-
-                debate.sendNewQuestion(new debate.Question('Does this test work ?', null, true));
-            });
-
-            it('invalid object', (done) => {
-                client.on('newQuestion', (questionObj) => {
-                    client.emit('answerOpenQuestion', {myFieldIsInvalid: 12}, (res) => {
-                        res.should.equal(false);
-                        done();
-                    });
-                });
-
-                debate.sendNewQuestion(new debate.Question('Does this test work ?', null, true));
-            });
-        });
-
-        afterEach(() => {
-            client.close();
-        });
-
-        after(() => {
-            admin.close();
-        });
-    });
-
     describe("Debate admin functions", () => {
         let client;
         let id;
